change to float

+++ recipe/models.py
@@ -77,7 +77,7 @@ class Recipe(models.Model):
 class Ingredient(models.Model):
     product = models.ForeignKey(Product, on_delete=models.CASCADE, related_name="product",null=True)
     recipe = models.ForeignKey(Recipe, models.CASCADE, null=True)
-    count = models.IntegerField(default=0)
+    count = models.FloatField(default=0)
 
     def __str__(self):
         return self.product.name
